requirements.txt
amqp==5.1.1
async-timeout==4.0.3
#backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cpe==1.2.1
dnspython==2.4.2
dotmap==1.3.30
eventlet==0.33.3
flower==2.0.1
free-proxy==1.1.1
furl==2.1.3
gevent==23.9.1
greenlet==2.0.2
humanize==4.8.0
idna==3.4
ifaddr==0.2.0
Jinja2==3.1.2
kombu==5.3.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
memray==1.9.1
orderedmultidict==1.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
Pygments==2.16.1
pyinstrument==4.5.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rich==13.5.3
rich-click==1.6.1
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
termcolor==2.3.0
tornado==6.3.3
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
validators==0.22.0
vine==5.0.0
wcwidth==0.2.6
xmltodict==0.13.0
